+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- support for relative module loading in starlark scripts.
+
 ### Breaking
 - removed deprecated kubernetes integration in favor of official kubernetes runner.
 - removed deprecated nomad integration in favor of official nomad runner.